the chopping off of trailing 00's was used previously to compensate for rjindaal padding that is no longer used.  While this shouldn't have caused a bug it is extraneous so I'm removing it
It only would cause a bug if d were divisible by 256 which
would mean that d is even.  that wouldn't make a lot of sense because d
is the multiplicative inverse of 0x10001 in cyclic ring of order
(p-1)(q-1).  Since (p-1)(q-1) is even, this means d isn't even.